feat(ui): reduce studies header height
